Переделал фигуры под пути и полигоны
