+++ ViewModels/FigureFabric.cs
@@ -16,9 +16,7 @@ class FigureMetadata
 }
 public interface IFigureCreator
 {
-    IReadOnlyCollection<string> PointParametersNames { get; }
-    IReadOnlyCollection<string> DoubleParametersNames { get; }
-    IFigure Create(Group parentGroup);
+    IFigure Create(Group parentGroup, IDictionary<string, Point> pointParams);
 }
 
 public static class FigureFabric
@@ -48,9 +46,9 @@ public static class FigureFabric
     }
 
     public static IEnumerable<string> AvailableFigures => info.AvailableFigures.Select(f => f.Metadata.Name);
-    public static IFigure CreateFigure(string FigureName, Group parentGroup)
+    public static IFigure CreateFigure(string FigureName, Group parentGroup, IDictionary<string, Point> pointParams)
     {
         return info.AvailableFigures.First(f => f.Metadata.Name == FigureName).Value
-            .Create(parentGroup);
+            .Create(parentGroup, pointParams);
     }
 }
